+++ ribs/archives/_archive_base.py
@@ -2,6 +2,7 @@
 from abc import ABC, abstractmethod
 
 import numpy as np
+from numpy_groupies import aggregate_nb as aggregate
 
 from ribs._utils import (check_batch_shape, check_finite, check_is_1d,
                          check_shape, np_scalar, validate_batch,
@@ -9,10 +10,6 @@ from ribs._utils import (check_batch_shape, check_finite, check_is_1d,
 from ribs.archives._archive_stats import ArchiveStats
 from ribs.archives._array_store import ArrayStore
 from ribs.archives._cqd_score_result import CQDScoreResult
-from ribs.archives._transforms import (batch_entries_with_threshold,
-                                       compute_best_index,
-                                       compute_objective_sum,
-                                       single_entry_with_threshold)
 
 _ARCHIVE_FIELDS = {"index", "solution", "objective", "measures", "threshold"}
 
@@ -93,10 +90,10 @@ class ArchiveBase(ABC):
         seed (int): Value to seed the random number generator. Set to None to
             avoid a fixed seed.
         dtype (str or data-type or dict): Data type of the solutions,
-            objectives, and measures. We only support ``"f"`` / ``np.float32``
-            and ``"d"`` / ``np.float64``. Alternatively, this can be a dict
-            specifying separate dtypes, of the form ``{"solution": <dtype>,
-            "objective": <dtype>, "measures": <dtype>}``.
+            objectives, and measures. This can be ``"f"`` / ``np.float32``,
+            ``"d"`` / ``np.float64``, or a dict specifying separate dtypes, of
+            the form ``{"solution": <dtype>, "objective": <dtype>, "measures":
+            <dtype>}``.
         extra_fields (dict): Description of extra fields of data that is stored
             next to elite data like solutions and objectives. The description is
             a dict mapping from a field name (str) to a tuple of ``(shape,
@@ -117,19 +114,19 @@ class ArchiveBase(ABC):
         ValueError: Invalid names in extra_fields.
     """
 
-    def __init__(self,
-                 *,
-                 solution_dim,
-                 cells,
-                 measure_dim,
-                 learning_rate=None,
-                 threshold_min=-np.inf,
-                 qd_score_offset=0.0,
-                 seed=None,
-                 dtype=np.float64,
-                 extra_fields=None):
-
-        self._seed = seed
+    def __init__(
+        self,
+        *,
+        solution_dim,
+        cells,
+        measure_dim,
+        learning_rate=None,
+        threshold_min=-np.inf,
+        qd_score_offset=0.0,
+        seed=None,
+        dtype=np.float64,
+        extra_fields=None,
+    ):
         self._rng = np.random.default_rng(seed)
         self._solution_dim = solution_dim
         self._measure_dim = measure_dim
@@ -376,6 +373,47 @@ class ArchiveBase(ABC):
                                dtype=self.dtypes["objective"]),
         )
 
+    @staticmethod
+    def _compute_thresholds(indices, objective, cur_threshold, learning_rate,
+                            dtype):
+        """Computes new thresholds with the CMA-MAE batch threshold update rule.
+
+        If entries in `indices` are duplicated, they receive the same threshold.
+        """
+        if len(indices) == 0:
+            return np.array([], dtype=dtype)
+
+        # Compute the number of objectives inserted into each cell. Note that we
+        # index with `indices` to place the counts at all relevant indices. For
+        # instance, if we had an array [1,2,3,1,5], we would end up with
+        # [2,1,1,2,1] (there are 2 1's, 1 2, 1 3, 2 1's, and 1 5).
+        #
+        # All objective_sizes should be > 0 since we only retrieve counts for
+        # indices in `indices`.
+        objective_sizes = aggregate(indices, 1, func="len",
+                                    fill_value=0)[indices]
+
+        # Compute the sum of the objectives inserted into each cell -- again, we
+        # index with `indices`.
+        objective_sums = aggregate(indices,
+                                   objective,
+                                   func="sum",
+                                   fill_value=np.nan)[indices]
+
+        # Update the threshold with the batch update rule from Fontaine 2023
+        # (https://arxiv.org/abs/2205.10752).
+        #
+        # Unlike in single_entry_with_threshold, we do not need to worry about
+        # cur_threshold having -np.inf here as a result of threshold_min being
+        # -np.inf. This is because the case with threshold_min = -np.inf is
+        # handled separately since we compute the new threshold based on the max
+        # objective in each cell in that case.
+        ratio = np_scalar(1.0 - learning_rate, dtype=dtype)**objective_sizes
+        new_threshold = (ratio * cur_threshold +
+                         (objective_sums / objective_sizes) * (1 - ratio))
+
+        return new_threshold
+
     def add(self, solution, objective, measures, **fields):
         """Inserts a batch of solutions into the archive.
 
@@ -396,9 +434,9 @@ class ArchiveBase(ABC):
         in the appendix of `Fontaine 2023 <https://arxiv.org/abs/2205.10752>`_.
 
         .. note:: The indices of all arguments should "correspond" to each
-            other, i.e. ``solution[i]``, ``objective[i]``,
-            ``measures[i]``, and should be the solution parameters,
-            objective, and measures for solution ``i``.
+            other, i.e., ``solution[i]``, ``objective[i]``, and ``measures[i]``
+            should be the solution parameters, objective, and measures for
+            solution ``i``.
 
         Args:
             solution (array-like): (batch_size, :attr:`solution_dim`) array of
@@ -431,13 +469,13 @@ class ArchiveBase(ABC):
 
               The alternative is to depend on the order of the solutions in the
               batch -- for example, if we have two solutions ``a`` and ``b``
-              which introduce the same new cell in the archive, ``a`` could be
+              that introduce the same new cell in the archive, ``a`` could be
               inserted first with status ``2``, and ``b`` could be inserted
               second with status ``1`` because it improves upon ``a``. However,
               our implementation does **not** do this.
 
               To convert statuses to a more semantic format, cast all statuses
-              to :class:`AddStatus` e.g. with ``[AddStatus(s) for s in
+              to :class:`AddStatus`, e.g., with ``[AddStatus(s) for s in
               add_info["status"]]``.
 
             - ``"value"`` (:class:`numpy.ndarray` of
@@ -447,14 +485,14 @@ class ArchiveBase(ABC):
               depends on the corresponding ``status`` and is identical to that
               in CMA-ME (`Fontaine 2020 <https://arxiv.org/abs/1912.02400>`_):
 
-              - ``0`` (not added): The value is the "negative improvement," i.e.
-                the objective of the solution passed in minus the objective of
-                the elite still in the archive (this value is negative because
-                the solution did not have a high enough objective to be added to
-                the archive).
+              - ``0`` (not added): The value is the "negative improvement,"
+                i.e., the objective of the solution passed in minus the
+                objective of the elite still in the archive (this value is
+                negative because the solution did not have a high enough
+                objective to be added to the archive).
               - ``1`` (improve existing cell): The value is the "improvement,"
-                i.e. the objective of the solution passed in minus the objective
-                of the elite previously in the archive.
+                i.e., the objective of the solution passed in minus the
+                objective of the elite previously in the archive.
               - ``2`` (new cell): The value is just the objective of the
                 solution.
 
@@ -478,26 +516,105 @@ class ArchiveBase(ABC):
             },
         )
 
-        add_info = self._store.add(
-            self.index_of(data["measures"]),
-            data,
-            {
-                "dtype": self.dtypes["threshold"],
-                "learning_rate": self._learning_rate,
-                "threshold_min": self._threshold_min,
-                "objective_sum": self._objective_sum,
-            },
-            [
-                batch_entries_with_threshold,
-                compute_objective_sum,
-                compute_best_index,
-            ],
-        )
+        # Delete these so that we only use the clean, validated data in `data`.
+        del solution, objective, measures, fields
 
-        objective_sum = add_info.pop("objective_sum")
-        best_index = add_info.pop("best_index")
-        if not np.all(add_info["status"] == 0):
-            self._stats_update(objective_sum, best_index)
+        # Information to return about the addition.
+        add_info = {}
+
+        # Retrieve indices of the archive cells.
+        indices = self.index_of(data["measures"])
+        batch_size = len(indices)
+
+        # Retrieve current data and thresholds. Unoccupied cells default to
+        # threshold_min.
+        cur_occupied, cur_data = self._store.retrieve(indices)
+        cur_threshold = cur_data["threshold"]
+        cur_threshold[~cur_occupied] = self.threshold_min
+
+        # Compute status -- arrays below are all boolean arrays of length
+        # batch_size.
+        #
+        # When we want CMA-ME behavior, the threshold defaults to -inf for new
+        # cells, which satisfies the condition for can_insert.
+        can_insert = data["objective"] > cur_threshold
+        is_new = can_insert & ~cur_occupied
+        improve_existing = can_insert & cur_occupied
+        add_info["status"] = np.zeros(batch_size, dtype=np.int32)
+        add_info["status"][is_new] = 2
+        add_info["status"][improve_existing] = 1
+
+        # If threshold_min is -inf, then we want CMA-ME behavior, which computes
+        # the improvement value of new solutions w.r.t zero. Otherwise, we
+        # compute improvement with respect to threshold_min.
+        cur_threshold[is_new] = (np_scalar(0.0, dtype=self.dtypes["threshold"])
+                                 if self.threshold_min == -np.inf else
+                                 self.threshold_min)
+        add_info["value"] = data["objective"] - cur_threshold
+
+        # Return early if we cannot insert anything -- continuing throws a
+        # ValueError in aggregate() since index[can_insert] would be empty.
+        if not np.any(can_insert):
+            return add_info
+
+        # Select all solutions that _can_ be inserted -- at this point, there
+        # are still conflicts in the insertions, e.g., multiple solutions can
+        # map to index 0.
+        indices = indices[can_insert]
+        data = {name: arr[can_insert] for name, arr in data.items()}
+        cur_threshold = cur_threshold[can_insert]
+
+        # Compute the new threshold associated with each entry.
+        if self.threshold_min == -np.inf:
+            # Regular archive behavior: thresholds are just the objectives.
+            new_threshold = data["objective"]
+        else:
+            # Batch threshold update described in Fontaine 2023
+            # (https://arxiv.org/abs/2205.10752). This computation is based on
+            # the mean objective of all solutions in the batch that could have
+            # been inserted into each cell.
+            new_threshold = self._compute_thresholds(indices, data["objective"],
+                                                     cur_threshold,
+                                                     self.learning_rate,
+                                                     self.dtypes["threshold"])
+
+        # Retrieve indices of solutions that _should_ be inserted into the
+        # archive. Currently, multiple solutions may be inserted at each archive
+        # index, but we only want to insert the maximum among these solutions.
+        # Thus, we obtain the argmax for each archive index.
+        #
+        # We use a fill_value of -1 to indicate archive indices that were not
+        # covered in the batch. Note that the length of archive_argmax is only
+        # max(indices), rather than the total number of grid cells. However,
+        # this is okay because we only need the indices of the solutions, which
+        # we store in should_insert.
+        #
+        # aggregate() always chooses the first item if there are ties, so the
+        # first elite will be inserted if there is a tie. See their default
+        # numpy implementation for more info:
+        # https://github.com/ml31415/numpy-groupies/blob/master/numpy_groupies/aggregate_numpy.py#L107
+        archive_argmax = aggregate(indices,
+                                   data["objective"],
+                                   func="argmax",
+                                   fill_value=-1)
+        should_insert = archive_argmax[archive_argmax != -1]
+
+        # Select only solutions that will be inserted into the archive.
+        indices = indices[should_insert]
+        data = {name: arr[should_insert] for name, arr in data.items()}
+        data["threshold"] = new_threshold[should_insert]
+
+        # Insert elites into the store.
+        self._store.add(indices, data)
+
+        # Compute statistics.
+        cur_objective = cur_data["objective"]
+        cur_objective[~cur_occupied] = 0.0
+        cur_objective = cur_objective[can_insert][should_insert]
+        objective_sum = (self._objective_sum +
+                         np.sum(data["objective"] - cur_objective))
+        best_index = indices[np.argmax(data["objective"])]
+        self._stats_update(objective_sum, best_index)
 
         return add_info
 
@@ -511,10 +628,9 @@ class ArchiveBase(ABC):
         also updated if the solution was inserted.
 
         .. note::
-            To make it more amenable to modifications, this method's
-            implementation is designed to be readable at the cost of
-            performance, e.g., none of its operations are modified. If you need
-            performance, we recommend using :meth:`add`.
+            This method is provided as an easier-to-understand implementation
+            that has less performance due to inserting only one solution at a
+            time. For better performance, see :meth:`add`.
 
         Args:
             solution (array-like): Parameters of the solution.
@@ -542,33 +658,84 @@ class ArchiveBase(ABC):
             },
         )
 
-        for name, arr in data.items():
-            data[name] = np.expand_dims(arr, axis=0)
+        # Delete these so that we only use the clean, validated data in `data`.
+        del solution, objective, measures, fields
 
-        add_info = self._store.add(
-            np.expand_dims(self.index_of_single(measures), axis=0),
-            data,
-            {
-                "dtype": self.dtypes["threshold"],
-                "learning_rate": self._learning_rate,
-                "threshold_min": self._threshold_min,
-                "objective_sum": self._objective_sum,
-            },
-            [
-                single_entry_with_threshold,
-                compute_objective_sum,
-                compute_best_index,
-            ],
-        )
+        # Information to return about the addition.
+        add_info = {}
 
-        objective_sum = add_info.pop("objective_sum")
-        best_index = add_info.pop("best_index")
+        # Identify the archive cell.
+        index = self.index_of_single(data["measures"])
 
-        for name, arr in add_info.items():
-            add_info[name] = arr[0]
+        # Retrieve current data of the cell.
+        cur_occupied, cur_data = self._store.retrieve([index])
+        cur_occupied = cur_occupied[0]
 
-        if add_info["status"]:
-            self._stats_update(objective_sum, best_index)
+        if cur_occupied:
+            # If the cell is currently occupied, the threshold comes from the
+            # current data of the elite in the cell.
+            cur_threshold = cur_data["threshold"][0]
+        else:
+            # If the cell is not currently occupied, the threshold needs special
+            # settings.
+            #
+            # If threshold_min is -inf, then we want CMA-ME behavior, which
+            # computes the improvement value with a threshold of zero for new
+            # solutions. Otherwise, we will set cur_threshold to threshold_min.
+            cur_threshold = (np_scalar(0.0, dtype=self.dtypes["threshold"])
+                             if self.threshold_min == -np.inf else
+                             self.threshold_min)
+
+        # Retrieve candidate objective.
+        objective = data["objective"]
+
+        # Compute status and threshold.
+        add_info["status"] = np.int32(0)  # NOT_ADDED
+
+        # Now we check whether a solution should be added to the archive. We use
+        # the addition rule from MAP-Elites (Fig. 2 of Mouret 2015
+        # https://arxiv.org/pdf/1504.04909.pdf), with modifications for CMA-MAE.
+
+        # This checks if a new solution is discovered in the archive. Note that
+        # regular MAP-Elites only checks `not cur_occupied`. CMA-MAE has an
+        # additional `threshold_min` that the objective must exceed for new
+        # cells. If CMA-MAE is not being used, then `threshold_min` is -np.inf,
+        # making this check identical to that of MAP-Elites.
+        is_new = not cur_occupied and self.threshold_min < objective
+
+        # This checks whether the solution improves an existing cell in the
+        # archive, i.e., whether it performs better than the current elite in
+        # this cell. Vanilla MAP-Elites compares to the objective of the cell's
+        # current elite. CMA-MAE compares to a threshold value that updates
+        # over time (i.e., cur_threshold). When learning_rate is set to 1.0 (the
+        # default value), we recover the same rule as in MAP-Elites because
+        # cur_threshold is equivalent to the objective of the solution in the
+        # cell.
+        improve_existing = cur_occupied and cur_threshold < objective
+
+        if is_new or improve_existing:
+            if improve_existing:
+                add_info["status"] = np.int32(1)  # IMPROVE_EXISTING
+            else:
+                add_info["status"] = np.int32(2)  # NEW
+
+            # This calculation works in the case where threshold_min is -inf
+            # because cur_threshold will be set to 0.0 instead.
+            data["threshold"] = [(cur_threshold * (1.0 - self.learning_rate) +
+                                  objective * self.learning_rate)]
+
+            self._store.add(index[None], {
+                name: np.expand_dims(arr, axis=0) for name, arr in data.items()
+            })
+
+            # Update stats.
+            cur_objective = (cur_data["objective"] if cur_occupied else
+                             np_scalar(0.0, self.dtypes["objective"]))
+            self._stats_update(self._objective_sum + objective - cur_objective,
+                               index)
+
+        # Value is the improvement over the current threshold (can be negative).
+        add_info["value"] = objective - cur_threshold
 
         return add_info