+++ tests/archives/array_store_test.py
@@ -77,8 +77,6 @@ def test_add_wrong_keys(store):
                 "measures": [[1.0, 2.0], [3.0, 4.0]],
                 # Missing `solution` key.
             },
-            {},  # Empty extra_args.
-            [],  # Empty transforms.
         )
 
 
@@ -91,8 +89,6 @@ def test_add_mismatch_indices(store):
                 "measures": [[1.0, 2.0], [3.0, 4.0]],
                 "solution": [np.zeros(10), np.ones(10)],
             },
-            {},  # Empty extra_args.
-            [],  # Empty transforms.
         )
 
 
@@ -105,8 +101,6 @@ def test_simple_add_retrieve_clear(store):
             "measures": [[1.0, 2.0], [3.0, 4.0]],
             "solution": [np.zeros(10), np.ones(10)],
         },
-        {},  # Empty extra_args.
-        [],  # Empty transforms.
     )
 
     assert len(store) == 2
@@ -137,8 +131,6 @@ def test_add_duplicate_indices(store):
             "measures": [[1.0, 2.0], [3.0, 4.0]],
             "solution": [np.zeros(10), np.ones(10)],
         },
-        {},  # Empty extra_args.
-        [],  # Empty transforms.
     )
 
     assert len(store) == 1
@@ -154,8 +146,6 @@ def test_dtypes(store):
             "measures": [[1.0, 2.0], [3.0, 4.0]],
             "solution": [np.zeros(10), np.ones(10)],
         },
-        {},  # Empty extra_args.
-        [],  # Empty transforms.
     )
 
     _, data = store.retrieve([5, 3])
@@ -176,8 +166,6 @@ def test_retrieve_duplicate_indices(store):
             "measures": [[3.0, 4.0]],
             "solution": [np.ones(10)],
         },
-        {},  # Empty extra_args.
-        [],  # Empty transforms.
     )
 
     occupied, data = store.retrieve([3, 3])
@@ -220,8 +208,6 @@ def test_retrieve(return_type, store):
             "measures": [[1.0, 2.0], [3.0, 4.0]],
             "solution": [np.zeros(10), np.ones(10)],
         },
-        {},  # Empty extra_args.
-        [],  # Empty transforms.
     )
 
     occupied, data = store.retrieve([5, 3], return_type=return_type)
@@ -279,8 +265,6 @@ def test_retrieve_custom_fields(store, return_type):
             "measures": [[1.0, 2.0], [3.0, 4.0]],
             "solution": [np.zeros(10), np.ones(10)],
         },
-        {},  # Empty extra_args.
-        [],  # Empty transforms.
     )
 
     occupied, data = store.retrieve([5, 3],
@@ -317,72 +301,12 @@ def test_retrieve_single_field(store):
             "measures": [[1.0, 2.0], [3.0, 4.0]],
             "solution": [np.zeros(10), np.ones(10)],
         },
-        {},  # Empty extra_args.
-        [],  # Empty transforms.
     )
 
     occupied, data = store.retrieve([5, 3], fields="objective")
 
     assert np.all(occupied == [True, True])
     assert np.all(data == [2.0, 1.0])
-
-
-def test_add_simple_transform(store):
-
-    def obj_meas(indices, new_data, add_info, extra_args, occupied, cur_data):
-        # pylint: disable = unused-argument
-        new_data["objective"] = np.sum(new_data["solution"], axis=1)
-        new_data["measures"] = np.asarray(new_data["solution"])[:, :2]
-        add_info.update(extra_args)
-        add_info["bar"] = 5
-        return indices, new_data, add_info
-
-    add_info = store.add(
-        [3, 5],
-        {
-            "solution": [np.ones(10), 2 * np.ones(10)],
-        },
-        {"foo": 4},
-        [obj_meas],
-    )
-
-    assert add_info == {"foo": 4, "bar": 5}
-
-    assert len(store) == 2
-    assert np.all(store.occupied == [0, 0, 0, 1, 0, 1, 0, 0, 0, 0])
-    assert np.all(np.sort(store.occupied_list) == [3, 5])
-
-    occupied, data = store.retrieve([3, 5])
-
-    assert np.all(occupied == [True, True])
-    assert data.keys() == set(["objective", "measures", "solution", "index"])
-    assert np.all(data["objective"] == [10.0, 20.0])
-    assert np.all(data["measures"] == [[1.0, 1.0], [2.0, 2.0]])
-    assert np.all(data["solution"] == [np.ones(10), 2 * np.ones(10)])
-    assert np.all(data["index"] == [3, 5])
-
-
-def test_add_empty_transform(store):
-    # new_data should be able to take on arbitrary values when no indices are
-    # returned, so we make it an empty dict here.
-    def empty(indices, new_data, add_info, extra_args, occupied, cur_data):
-        # pylint: disable = unused-argument
-        return [], {}, {}
-
-    add_info = store.add(
-        [3, 5],
-        {
-            "solution": [np.ones(10), 2 * np.ones(10)],
-        },
-        {"foo": 4},
-        [empty],
-    )
-
-    assert add_info == {}
-
-    assert len(store) == 0
-    assert np.all(~store.occupied)
-    assert len(store.occupied_list) == 0
 
 
 def test_resize_bad_capacity(store):
@@ -398,8 +322,6 @@ def test_resize_to_double_capacity(store):
             "measures": [[1.0, 2.0], [3.0, 4.0]],
             "solution": [np.zeros(10), np.ones(10)],
         },
-        {},  # Empty extra_args.
-        [],  # Empty transforms.
     )
 
     store.resize(store.capacity * 2)
@@ -421,8 +343,6 @@ def test_as_raw_dict(store):
             "measures": [[1.0, 2.0], [3.0, 4.0]],
             "solution": [np.zeros(10), np.ones(10)],
         },
-        {},  # Empty extra_args.
-        [],  # Empty transforms.
     )
 
     d = store.as_raw_dict()
@@ -462,8 +382,6 @@ def test_from_raw_dict(store):
             "measures": [[1.0, 2.0], [3.0, 4.0]],
             "solution": [np.zeros(10), np.ones(10)],
         },
-        {},  # Empty extra_args.
-        [],  # Empty transforms.
     )
 
     new_store = ArrayStore.from_raw_dict(store.as_raw_dict())
@@ -490,8 +408,6 @@ def test_data(store):
             "measures": [[1.0, 2.0], [3.0, 4.0]],
             "solution": [np.zeros(10), np.ones(10)],
         },
-        {},  # Empty extra_args.
-        [],  # Empty transforms.
     )
 
     d = store.data()
@@ -520,8 +436,6 @@ def test_data_with_tuple_return_type(store):
             "measures": [[1.0, 2.0], [3.0, 4.0]],
             "solution": [np.zeros(10), np.ones(10)],
         },
-        {},  # Empty extra_args.
-        [],  # Empty transforms.
     )
 
     d = store.data(return_type="tuple")
@@ -550,8 +464,6 @@ def test_data_with_pandas_return_type(store):
             "measures": [[1.0, 2.0], [3.0, 4.0]],
             "solution": [np.zeros(10), np.ones(10)],
         },
-        {},  # Empty extra_args.
-        [],  # Empty transforms.
     )
 
     df = store.data(return_type="pandas")
@@ -591,8 +503,6 @@ def test_iteration(store):
             "measures": [[1.0, 2.0]],
             "solution": [np.zeros(10)],
         },
-        {},  # Empty extra_args.
-        [],  # Empty transforms.
     )
 
     for entry in store:
@@ -612,8 +522,6 @@ def test_add_during_iteration(store):
             "measures": [[1.0, 2.0]],
             "solution": [np.zeros(10)],
         },
-        {},  # Empty extra_args.
-        [],  # Empty transforms.
     )
 
     # Even with just one entry, adding during iteration should still raise an
@@ -627,8 +535,6 @@ def test_add_during_iteration(store):
                     "measures": [[3.0, 4.0]],
                     "solution": [np.ones(10)],
                 },
-                {},  # Empty extra_args.
-                [],  # Empty transforms.
             )
 
 
@@ -640,8 +546,6 @@ def test_clear_during_iteration(store):
             "measures": [[1.0, 2.0]],
             "solution": [np.zeros(10)],
         },
-        {},  # Empty extra_args.
-        [],  # Empty transforms.
     )
 
     with pytest.raises(RuntimeError):
@@ -657,8 +561,6 @@ def test_clear_and_add_during_iteration(store):
             "measures": [[1.0, 2.0]],
             "solution": [np.zeros(10)],
         },
-        {},  # Empty extra_args.
-        [],  # Empty transforms.
     )
 
     with pytest.raises(RuntimeError):
@@ -671,6 +573,4 @@ def test_clear_and_add_during_iteration(store):
                     "measures": [[3.0, 4.0]],
                     "solution": [np.ones(10)],
                 },
-                {},  # Empty extra_args.
-                [],  # Empty transforms.
             )